requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
python-multipart==0.0.9
pydantic[email]==2.9.2
pydantic-settings==2.5.2
alembic==1.13.3
python-socketio[asyncio]==5.11.0

# AI/RAG Dependencies
google-generativeai==0.7.0
chromadb==0.5.0
langchain==0.3.0
langchain-google-genai==2.0.0
langchain-community==0.3.0
beautifulsoup4==4.12.3
pypdf==4.0.0
httpx==0.27.0